fixed getPointer & drag functions for touch screen
getPointer and drag were not setting position when using a touch device.

This quick fix may not work properly when using multitouch but I didn't saw any problem on chrome emulation and on my phone.

There is multiple ways to get the touch event's position and I might not use the best one. Maybe someone with experience in the matter could improve this solution?